fix: decouple note management with canvas
final implementation of NoteContainer
